New (Better) grass to Water tiles... (read description for longer rant)
I wasn't sure if anyone was currently using the ones I had dumped into the repo originally that, as pointed out, were not very good or thought out. So I put these in their own folder (I didn't want to cause conflict with anything, and the reality is that there are very different number of tiles for this batch vs the previous batch). I will endeavor to get the sand to water tiles done by this afternoon.